Remove the usage of CD Key and related infrastructure
This was a secret supplied when purchasing the product and only one player
could use the secret to play at a time. A separate authorization server was
kept track of users globally and both the client and server had to talk to
the auth server prior to establishing a connection. This infrastructure
will not be used going forward with a standalone game.

Removal of this infrastructure also removed the sv_strictAuth cvar as well
as the console commands banUser and banClient