Fix issue with multiple item pickup events playing
This was causing horrible audio, as multiple item pickup events were all playing at the same time. This was because the predict player state was being called once per eye, and by the time it came to the right eye it had cleared the buffer and it would just allow server events to get played again. The actual fix is in cg_view.c, line 994, where we only predict if the eye being rendered is the left.