Format the source code using clang-format
There was no dominant style across the entire code base so a relatively
arbitrary decision to adopt the builting style WebKit but with #include
sorting disabled (as the headers are not well defined and this would lead
to compilation errors.

Formatting source code is boring and leads to pointless debates.
A quotation from "Rust Style Guide":

> Formatting code is a mostly mechanical task which takes both time and
> mental effort. By using an automatic formatting tool, a programmer is
> relieved of this task and can concentrate on more important things.
> Furthermore, by sticking to an established style guide (such as this
> one), programmers don't need to formulate ad hoc style rules, nor do
> they need to debate with other programmers what style rules should be
> used, saving time, communication overhead, and mental energy.