Replace strcpy with strncpyz where feasible
The intent is to remove potential crashes by range checking
string operations. Some places where strcpy was replaced have
no chance of overflowing the destination variable as the src
variable is a literal and the destination is a fixed size array
of greater capacity.

The decision was made to still move to strncpyz as the perfromance
impact is insignificant and it does reduce the cognitive load on
reader if all operations are treated consistently.